#56 feat 채용/면접 쿼리메소드 구현

+++ src/main/java/com/errorCode/pandaOffice/recruitment/domain/repository/InterviewScheduleRepository.java
@@ -3,5 +3,5 @@ package com.errorCode.pandaOffice.recruitment.domain.repository;
 import com.errorCode.pandaOffice.recruitment.domain.entity.InterviewSchedule;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InterviewScheduleRepository extends JpaRepository<InterviewSchedule, Integer> {
+public interface InterviewScheduleRepository extends JpaRepository<InterviewSchedule, Long> {
 }
